+++ atlasdb-client/src/main/java/com/palantir/atlasdb/cache/TimestampCache.java
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright 2018 Palantir Technologies Inc. All rights reserved.
+ * (c) Copyright 2019 Palantir Technologies Inc. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,66 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.palantir.atlasdb.cache;
 
-import java.util.function.Supplier;
+package com.palantir.atlasdb.cache;
 
 import javax.annotation.Nullable;
 
-import com.codahale.metrics.MetricRegistry;
-import com.github.benmanes.caffeine.cache.Cache;
-import com.github.benmanes.caffeine.cache.Caffeine;
-import com.github.benmanes.caffeine.cache.Policy;
-import com.google.common.annotations.VisibleForTesting;
-import com.palantir.atlasdb.util.AtlasDbMetrics;
-
-public class TimestampCache {
-    private final Supplier<Long> size;
-
-    private final Cache<Long, Long> startToCommitTimestampCache;
-    private final Policy.Eviction<Long, Long> evictionPolicy;
-
-    @VisibleForTesting
-    static Cache<Long, Long> createCache(long size) {
-        return Caffeine.newBuilder()
-                .maximumSize(size)
-                .recordStats()
-                .build();
-    }
-
-    public TimestampCache(MetricRegistry metricRegistry, Supplier<Long> size) {
-        this.size = size;
-        startToCommitTimestampCache = createCache(size.get());
-        evictionPolicy = startToCommitTimestampCache.policy().eviction().get();
-        AtlasDbMetrics.registerCache(metricRegistry, startToCommitTimestampCache,
-                MetricRegistry.name(TimestampCache.class, "startToCommitTimestamp"));
-    }
-
-    @VisibleForTesting
-    TimestampCache(Cache<Long, Long> cache) {
-        this.evictionPolicy = cache.policy().eviction().get();
-        this.size = evictionPolicy::getMaximum;
-        this.startToCommitTimestampCache = cache;
-    }
-
+public interface TimestampCache {
     /**
-     * Returns null if not present.
-     *
-     * @param startTimestamp transaction start timestamp
-     * @return commit timestamp for the specified transaction start timestamp if present in cache, otherwise null
+     * Clear all values from the cache.
      */
-    @Nullable
-    public Long getCommitTimestampIfPresent(Long startTimestamp) {
-        resizeIfNecessary();
-        return startToCommitTimestampCache.getIfPresent(startTimestamp);
-    }
-
-    private void resizeIfNecessary() {
-        if (evictionPolicy.getMaximum() != size.get()) {
-            evictionPolicy.setMaximum(size.get());
-        }
-    }
-
+    void clear();
     /**
      * Be very careful to only insert timestamps here that are already present in the backing store,
      * effectively using the timestamp table as existing concurrency control for who wins a commit.
@@ -80,18 +30,13 @@ public class TimestampCache {
      * @param startTimestamp transaction start timestamp
      * @param commitTimestamp transaction commit timestamp
      */
-    public void putAlreadyCommittedTransaction(Long startTimestamp, Long commitTimestamp) {
-        startToCommitTimestampCache.put(startTimestamp, commitTimestamp);
-    }
-
+    void putAlreadyCommittedTransaction(Long startTimestamp, Long commitTimestamp);
     /**
-     * Clear all values from the cache.
+     * Returns null if not present.
+     *
+     * @param startTimestamp transaction start timestamp
+     * @return commit timestamp for the specified transaction start timestamp if present in cache, otherwise null
      */
-    public void clear() {
-        startToCommitTimestampCache.invalidateAll();
-    }
-
-    public static TimestampCache createForTests() {
-        return new TimestampCache(new MetricRegistry(), () -> 1000L);
-    }
+    @Nullable
+    Long getCommitTimestampIfPresent(Long startTimestamp);
 }
